refactor: change id type to int

+++ src/main/kotlin/com/dms/messages/repository/MessagesRepository.kt
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface MessagesRepository : JpaRepository<Message, Long> {
+interface MessagesRepository : JpaRepository<Message, Int> {
 }
